// src/types/responseChannels.ts
export interface ResponseChannels {
  kind: string
  etag: string
  pageInfo: PageInfo
  items: ResponseChannelsItems[]
}

export interface ResponseChannelsItems {
  kind: string
  etag: string
  id: string
  snippet: Snippet
  statistics: Statistics
}

interface Statistics {
  viewCount: string
  subscriberCount: string
  hiddenSubscriberCount: boolean
  videoCount: string
}

interface Snippet {
  title: string
  description: string
  customUrl: string
  publishedAt: string
  thumbnails: Thumbnails
  localized: Localized
  country: string
}

interface Localized {
  title: string
  description: string
}

interface Thumbnails {
  default: Default
  medium: Default
  high: Default
}

interface Default {
  url: string
  width: number
  height: number
}

interface PageInfo {
  totalResults: number
  resultsPerPage: number
}